pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slm-agent"
version = "2.3.0"
description = "Agentic bug-bounty SLM that runs 100% local — from phones to workstations"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "DXN10DAY"}]
keywords = ["bug-bounty", "agent", "llm", "security", "termux", "local-ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
]
dependencies = [
    "typer>=0.9",
    "httpx>=0.25",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "pyyaml>=6.0",
    "sqlite-utils>=3.35",
    "beautifulsoup4>=4.12",
    "duckduckgo-search>=4.0",
    "tomli>=2.0; python_version<'3.11'",
]

[project.optional-dependencies]
tui = ["textual>=0.40"]
snowflake = ["snowflake-connector-python>=3.5"]
vault = ["cryptography>=42.0"]
discord = ["discord.py>=2.3", "cryptography>=42.0"]
cloud = ["cryptography>=42.0"]
train = [
    "torch>=2.1",
    "transformers>=4.36",
    "peft>=0.7",
    "trl>=0.7",
    "datasets>=2.16",
    "accelerate>=0.25",
]
dev = [
    "ruff>=0.3",
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
all = [
    "textual>=0.40",
    "cryptography>=42.0",
    "discord.py>=2.3",
    "snowflake-connector-python>=3.5",
]

[project.urls]
Homepage = "https://github.com/DXN1-termux/BUGHUNTER-AI"
Issues = "https://github.com/DXN1-termux/BUGHUNTER-AI/issues"
Discussions = "https://github.com/DXN1-termux/BUGHUNTER-AI/discussions"

[project.scripts]
slm = "slm.cli:app"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["slm*"]

[tool.setuptools.package-data]
"slm" = ["core/*.yaml", "prompts/*.md"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "S", "B"]
ignore = ["S603", "S607", "S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
